# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News-aggregator API Documentation",
        "description": "Swagger OpenAPI Description for case study of news aggregator api",
        "contact": {
            "email": "iqbal.shaheen0101@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "operationId": "93ded9e31119e6bbb22455c72ef97f19",
                "responses": {
                    "200": {
                        "description": "returns all articles"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/article/{id}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "operationId": "cc7a17f59973033eb40d6b13de411848",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Article ID",
                        "example": "10"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "searched  articles"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/articles/search": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "operationId": "b2a37f58fc1f822aac6033aeca788724",
                "parameters": [
                    {
                        "name": "source",
                        "in": "query",
                        "description": "source",
                        "example": "bbc-news"
                    },
                    {
                        "name": "published_at",
                        "in": "query",
                        "description": "published_at",
                        "example": "2023-12-03"
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "category",
                        "example": "sport"
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "description": "keywords",
                        "example": "Olympics 2024"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "searched  articles"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Register User",
                "operationId": "482fd8cbb98dc3225fa0e7ac40c1abe6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "User name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "User email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "set password",
                                        "type": "string"
                                    },
                                    "c_password": {
                                        "description": "Confirm password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User Registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login User",
                "operationId": "afd167661891a2c4c1b49a04643b9223",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "User email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User logedin successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Logout User",
                "operationId": "9872a1ded930c0faa99c7f184b4cd654",
                "responses": {
                    "200": {
                        "description": "User loged out successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/get-user-preferences": {
            "get": {
                "tags": [
                    "User Preference"
                ],
                "summary": "Get user preferences",
                "operationId": "a8c7163fa925040498784a5c619792d6",
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/set-user-preferences": {
            "post": {
                "tags": [
                    "User Preference"
                ],
                "summary": "Store user preferences",
                "operationId": "1abc59535cf4dbbdc141aa8e16db9a2b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "preferences[authors][0]": {
                                        "description": "Author 1",
                                        "type": "string"
                                    },
                                    "preferences[authors][1]": {
                                        "description": "Author 2",
                                        "type": "string"
                                    },
                                    "preferences[categories][0]": {
                                        "description": "sport",
                                        "type": "string"
                                    },
                                    "preferences[categories][1]": {
                                        "description": "art",
                                        "type": "string"
                                    },
                                    "preferences[sources][0]": {
                                        "description": "bbc-news",
                                        "type": "string"
                                    },
                                    "preferences[sources][1]": {
                                        "description": "bbc-news",
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Preferences saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/get-user-feeds": {
            "get": {
                "tags": [
                    "User Preference"
                ],
                "summary": "Get user feeds based on preferences",
                "operationId": "502bbf0c8f52a3ca7175f1dca7d84b29",
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "tags": [
        {
            "name": "Articles",
            "description": "Articles"
        },
        {
            "name": "User",
            "description": "User"
        },
        {
            "name": "User Preference",
            "description": "User Preference"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "A short description for security scheme",
                "name": "token",
                "in": "header",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}